+++ wallet/src/account/mod.rs
@@ -64,7 +64,7 @@ use crypto::{
     },
     vrf::VRFPublicKey,
 };
-use mempool::FeeRate;
+use mempool::{FeeRate, MempoolConfig};
 use serialization::hex_encoded::HexEncoded;
 use utils::{ensure, iter::CloneableExactSizeIterator};
 use wallet_storage::{
@@ -215,6 +215,17 @@ impl<K: AccountKeyChains> Account<K> {
     }
 
     // Note: the default selection algo depends on whether input_utxos are empty.
+    // TODO: utxo selection algorithm must be improved, so that it avoids creating transaction
+    // clusters that violate the corresponding mempool limits. Bitcoin's `AutomaticCoinSelection`
+    // may be used as an inspiration, e.g. we may want to:
+    // *) prefer confirmed utxos is general;
+    // *) prefer utxos that have fewer unconfirmed dependencies;
+    // *) take into account the number of confirmations, especially for utxos coming from a
+    //    different wallet.
+    // Also note that at the moment of writing this mempool only considers utxo-based relations to
+    // determine tx clusters. But this will be fixed in the (hopefully near) future, and all
+    // dependencies will be considered (e.g. the account-based ones and things like order creation
+    // vs order commands). The new algorithm will have to take this into account as well.
     #[allow(clippy::too_many_arguments)]
     pub fn select_inputs_for_send_request(
         &mut self,
@@ -226,6 +237,7 @@ impl<K: AccountKeyChains> Account<K> {
         median_time: BlockTimestamp,
         fee_rates: CurrentFeeRate,
         order_info: Option<BTreeMap<OrderId, &RpcOrderInfo>>,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         // TODO: allow to pay fees with different currency?
         let pay_fee_with_currency = Currency::Coin;
@@ -314,6 +326,17 @@ impl<K: AccountKeyChains> Account<K> {
             }
         };
 
+        // The maximum possible size for a tx. ChainConfig::max_tx_size_for_mempool() is the
+        // absolute limit, but a transaction cannot be bigger than the cluster size limit.
+        // TODO: currently this value is only passed to select_coins, but the actual tx size
+        // may become bigger after change is added (also, select_coins will ignore the passed value
+        // in the `CoinSelectionAlgo::UsePreselected` case). This should probably be fixed as a part
+        // of a general utxo selection revamp, see the TODO at the top of this function.
+        let max_tx_size = std::cmp::min(
+            self.chain_config.max_tx_size_for_mempool(),
+            *mempool_config.max_cluster_size_bytes,
+        );
+
         let mut selected_inputs: BTreeMap<_, _> = output_currency_amounts
             .iter()
             .map(|(currency, output_amount)| -> WalletResult<_> {
@@ -338,13 +361,7 @@ impl<K: AccountKeyChains> Account<K> {
                     // when we allow paying fees with different currency
                     Amount::ZERO,
                     selection_algo,
-                    // FIXME: Get mempool config from the node and use cluster size as the limit
-                    // for the size of 1 tx (or, better, take the minimum with chain_config.max_tx_size_for_mempool,
-                    // just in case).
-                    // Same below.
-                    // Create an issue to make the wallet take the cluster limit into account when
-                    // selecting utxos.
-                    self.chain_config.max_tx_size_for_mempool(),
+                    max_tx_size,
                 )?;
 
                 total_tx_size += selection_result.get_weight();
@@ -426,7 +443,7 @@ impl<K: AccountKeyChains> Account<K> {
             PayFee::PayFeeWithThisCurrency,
             cost_of_change,
             selection_algo,
-            self.chain_config.max_tx_size_for_mempool(),
+            max_tx_size,
         )?;
 
         let mut selection_result = selection_result.add_change(
@@ -941,6 +958,7 @@ impl<K: AccountKeyChains> Account<K> {
         htlc: HashedTimelockContract,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let output = TxOutput::Htlc(output_value, Box::new(htlc));
         let request = SendRequest::new().with_outputs([output]);
@@ -954,6 +972,7 @@ impl<K: AccountKeyChains> Account<K> {
             median_time,
             fee_rate,
             None,
+            mempool_config,
         )
     }
 
@@ -965,6 +984,7 @@ impl<K: AccountKeyChains> Account<K> {
         conclude_address: Address<Destination>,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let order_data =
             common::chain::OrderData::new(conclude_address.into_object(), ask_value, give_value);
@@ -980,6 +1000,7 @@ impl<K: AccountKeyChains> Account<K> {
             median_time,
             fee_rate,
             None,
+            mempool_config,
         )
     }
 
@@ -991,6 +1012,7 @@ impl<K: AccountKeyChains> Account<K> {
         output_address: Option<Destination>,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let output_destination = if let Some(dest) = output_address {
             dest
@@ -1049,6 +1071,7 @@ impl<K: AccountKeyChains> Account<K> {
             median_time,
             fee_rate,
             Some(BTreeMap::from_iter([(order_id, &order_info)])),
+            mempool_config,
         )
     }
 
@@ -1062,6 +1085,7 @@ impl<K: AccountKeyChains> Account<K> {
         output_address: Option<Destination>,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let output_destination = if let Some(dest) = output_address {
             dest
@@ -1144,6 +1168,7 @@ impl<K: AccountKeyChains> Account<K> {
             median_time,
             fee_rate,
             Some(BTreeMap::from_iter([(order_id, &order_info)])),
+            mempool_config,
         )
     }
 
@@ -1154,6 +1179,7 @@ impl<K: AccountKeyChains> Account<K> {
         order_info: RpcOrderInfo,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let request = SendRequest::new().with_inputs_and_destinations([(
             TxInput::OrderAccountCommand(OrderAccountCommand::FreezeOrder(order_id)),
@@ -1169,6 +1195,7 @@ impl<K: AccountKeyChains> Account<K> {
             median_time,
             fee_rate,
             Some(BTreeMap::from_iter([(order_id, &order_info)])),
+            mempool_config,
         )
     }
 
@@ -1180,6 +1207,7 @@ impl<K: AccountKeyChains> Account<K> {
         htlc_secret: Option<HtlcSecret>,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let utxo = self
             .output_cache
@@ -1298,6 +1326,7 @@ impl<K: AccountKeyChains> Account<K> {
                 median_time,
                 fee_rate,
                 None,
+                mempool_config,
             )
         }
     }
@@ -1308,6 +1337,7 @@ impl<K: AccountKeyChains> Account<K> {
         nft_issue_arguments: IssueNftArguments,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let nft_issuance = NftIssuanceV0 {
             metadata: nft_issue_arguments.metadata,
@@ -1333,6 +1363,7 @@ impl<K: AccountKeyChains> Account<K> {
             median_time,
             fee_rate,
             None,
+            mempool_config,
         )?;
 
         let new_token_id = make_token_id(
@@ -1375,6 +1406,7 @@ impl<K: AccountKeyChains> Account<K> {
         amount: Amount,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let token_id = token_info.token_id();
         let outputs = make_mint_token_outputs(token_id, amount, address);
@@ -1392,6 +1424,7 @@ impl<K: AccountKeyChains> Account<K> {
             db_tx,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 
@@ -1402,6 +1435,7 @@ impl<K: AccountKeyChains> Account<K> {
         amount: Amount,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let token_id = token_info.token_id();
         let outputs = make_unmint_token_outputs(token_id, amount);
@@ -1419,6 +1453,7 @@ impl<K: AccountKeyChains> Account<K> {
             db_tx,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 
@@ -1428,6 +1463,7 @@ impl<K: AccountKeyChains> Account<K> {
         token_info: &UnconfirmedTokenInfo,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let token_id = token_info.token_id();
         token_info.check_can_lock()?;
@@ -1443,6 +1479,7 @@ impl<K: AccountKeyChains> Account<K> {
             db_tx,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 
@@ -1453,6 +1490,7 @@ impl<K: AccountKeyChains> Account<K> {
         is_token_unfreezable: IsTokenUnfreezable,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         token_info.check_can_freeze()?;
 
@@ -1470,6 +1508,7 @@ impl<K: AccountKeyChains> Account<K> {
             db_tx,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 
@@ -1479,6 +1518,7 @@ impl<K: AccountKeyChains> Account<K> {
         token_info: &UnconfirmedTokenInfo,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         token_info.check_can_unfreeze()?;
 
@@ -1494,6 +1534,7 @@ impl<K: AccountKeyChains> Account<K> {
             db_tx,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 
@@ -1504,6 +1545,7 @@ impl<K: AccountKeyChains> Account<K> {
         address: Address<Destination>,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let new_authority = address.into_object();
 
@@ -1521,6 +1563,7 @@ impl<K: AccountKeyChains> Account<K> {
             db_tx,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 
@@ -1531,6 +1574,7 @@ impl<K: AccountKeyChains> Account<K> {
         metadata_uri: Vec<u8>,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let nonce = token_info.get_next_nonce()?;
         let tx_input = TxInput::AccountCommand(
@@ -1546,6 +1590,7 @@ impl<K: AccountKeyChains> Account<K> {
             db_tx,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 
@@ -1557,6 +1602,7 @@ impl<K: AccountKeyChains> Account<K> {
         db_tx: &mut impl WalletStorageWriteLocked,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> Result<SendRequest, WalletError> {
         let request = SendRequest::new()
             .with_outputs(outputs)
@@ -1571,6 +1617,7 @@ impl<K: AccountKeyChains> Account<K> {
             median_time,
             fee_rate,
             None,
+            mempool_config,
         )
     }
 
@@ -1580,6 +1627,7 @@ impl<K: AccountKeyChains> Account<K> {
         mut stake_pool_arguments: StakePoolCreationArguments,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let Some(vrf_public_key) = stake_pool_arguments.vrf_public_key.take() else {
             return Err(WalletError::VrfKeyMustBeProvided);
@@ -1591,6 +1639,7 @@ impl<K: AccountKeyChains> Account<K> {
             vrf_public_key,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 
@@ -1601,6 +1650,7 @@ impl<K: AccountKeyChains> Account<K> {
         vrf_public_key: VRFPublicKey,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> Result<SendRequest, WalletError> {
         let staker = match stake_pool_arguments.staker_key {
             Some(staker) => match staker {
@@ -1645,6 +1695,7 @@ impl<K: AccountKeyChains> Account<K> {
             median_time,
             fee_rate,
             None,
+            mempool_config,
         )?;
 
         let new_pool_id = common::chain::make_pool_id(request.inputs())?;
@@ -2626,6 +2677,7 @@ impl<K: AccountKeyChains + VRFAccountKeyChains> Account<K> {
         mut stake_pool_arguments: StakePoolCreationArguments,
         median_time: BlockTimestamp,
         fee_rate: CurrentFeeRate,
+        mempool_config: &MempoolConfig,
     ) -> WalletResult<SendRequest> {
         let vrf_public_key = match stake_pool_arguments.vrf_public_key.take() {
             Some(vrf_public_key) => vrf_public_key,
@@ -2637,6 +2689,7 @@ impl<K: AccountKeyChains + VRFAccountKeyChains> Account<K> {
             vrf_public_key,
             median_time,
             fee_rate,
+            mempool_config,
         )
     }
 }